+++ demos/chinook/src/main/java/is/codion/demos/chinook/domain/api/Chinook.java
@@ -286,7 +286,7 @@ public interface Chinook {
 		private static final long serialVersionUID = 1;
 
 		@Override
-		public BigDecimal get(SourceValues values) {
+		public BigDecimal from(SourceValues values) {
 			Integer quantity = values.get(InvoiceLine.QUANTITY);
 			BigDecimal unitPrice = values.get(InvoiceLine.UNITPRICE);
 			if (unitPrice == null || quantity == null) {